backend/requirements.txt
Flask>=2.0
Flask-Cors
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
gunicorn
psycopg2-binary
pytest
pytest-flask
python-dotenv
pyJWT
